feat: paper sturcture init

+++ internal/models/paper/paper.go
@@ -4,6 +4,6 @@ import "gorm.io/gorm"
 
 type Paper struct {
 	gorm.Model
-	Title     string  `gorm:"column:title;comment:'试卷标题'" json:"title"`
-	Questions []int32 `gorm:"column:questions;type:json;comment:'试卷包含题目ID'" json:"questions"`
+	Title     string `gorm:"column:title;comment:'试卷标题'" json:"title"`
+	Questions []byte `gorm:"column:questions;type:json;comment:'试卷包含题目ID'" json:"questions"`
 }
